+++ CHANGES.md
@@ -1,3 +1,9 @@
+[unreleased]
+
+#### 1.1.0 / 2021-12-18
+* fix to allow members to edit Los Robles data
+* improved handling of US phone numbers
+
 #### 1.0.0
 * refactor and general cleanup
 * add custom user fields for street number, phone number, and emergency contact